GEOMESA-766 Separate WPS processes so that users don't have to deploy the gs-wps plugin when installing the geomesa plugin
