+++ geomesa-assemble/pom.xml
@@ -58,12 +58,28 @@
         </dependency>
         <dependency>
             <groupId>org.locationtech.geomesa</groupId>
+            <artifactId>geomesa-wps-accumulo1.5</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.locationtech.geomesa</groupId>
             <artifactId>geomesa-distributed-runtime-accumulo1.5</artifactId>
         </dependency>
         <dependency>
             <groupId>org.locationtech.geomesa</groupId>
             <artifactId>geomesa-plugin-accumulo1.5</artifactId>
             <classifier>geoserver-plugin</classifier>
+        </dependency>
+        <dependency>
+            <groupId>org.locationtech.geomesa</groupId>
+            <artifactId>geomesa-web-core-accumulo1.5</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.locationtech.geomesa</groupId>
+            <artifactId>geomesa-web-csv-accumulo1.5</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.locationtech.geomesa</groupId>
+            <artifactId>geomesa-web-data-accumulo1.5</artifactId>
         </dependency>
         <dependency>
             <groupId>org.locationtech.geomesa</groupId>
