--- geomesa-plugin/src/main/java/org/locationtech/geomesa/plugin/wps/GeomesaProcess.java
@@ -1,4 +1,0 @@
-package org.locationtech.geomesa.plugin.wps;
-
-public interface GeomesaProcess {
-}
