--- geomesa-plugin/src/main/resources/META-INF/services/org.geotools.process.ProcessFactory
@@ -1,2 +1,0 @@
-org.locationtech.geomesa.plugin.process.ProcessFactory
-
