+++ geomesa-process/src/main/java/org/locationtech/geomesa/process/GeomesaProcess.java
@@ -1,0 +1,4 @@
+package org.locationtech.geomesa.process;
+
+public interface GeomesaProcess {
+}
