+++ geomesa-process/src/main/resources/META-INF/services/org.geotools.process.ProcessFactory
@@ -1,0 +1,2 @@
+org.locationtech.geomesa.process.GeomesaProcessFactory
+
