+++ geomesa-process/src/main/resources/applicationContext.xml
@@ -1,0 +1,33 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+  ~ Copyright 2014 Commonwealth Computer Research, Inc.
+  ~
+  ~ Licensed under the Apache License, Version 2.0 (the "License");
+  ~ you may not use this file except in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~ http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  -->
+
+<beans xmlns="http://www.springframework.org/schema/beans"
+       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+       xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd">
+
+
+    <bean id="geomesaProcessFactory" class="GeomesaProcessFactory">
+        <constructor-arg index="0" value="Geomesa Processes"/>
+        <constructor-arg index="1" value="geomesa"/>
+        <constructor-arg index="2" value="GeomesaProcess"/>
+    </bean>
+
+    <bean id="geomesaImportProcess" class="org.locationtech.geomesa.process.ImportProcess">
+        <constructor-arg index="0" ref="catalog"/>
+    </bean>
+
+</beans>
